+++ src/converter/computed-style-snapshot.browser.js
@@ -27,6 +27,7 @@
     "rowGap",
     "textAlign",
     "fontSize",
+    "lineHeight",
     "fontWeight",
     "fontStyle",
     "textTransform",